+++ src/dmkern.jl
@@ -48,7 +48,7 @@ fit!(model0, X, y)
 @head T = transf(model0, X)
 n, p = size(T)
 
-#### Probability density in the FDA score space (2D)
+#### Probability density in the FDA score space (2-D)
 
 model = dmkern()
 fit!(model, T)